Feat: 책 기록 삭제 API 코드 수정

+++ core/src/main/java/org/bookwoori/core/domain/record/service/RecordService.java
@@ -53,6 +53,9 @@ public class RecordService {
     }
 
     public void deleteRecord(Long recordId) {
+        if (!recordRepository.existsById(recordId)) {
+            throw new CustomException(ErrorCode.RECORD_NOT_FOUND);
+        }
         recordRepository.deleteById(recordId);
     }
